Always copy test plugins
When running `mvn clean install` the elasticsearch plugins are not copied to the `target` directory.
The problem is that when running `mvn clean`, the `clean` is ran after the evaluation of the maven profiles.

Because the file was existing, the profile is not activated.

This commit always copies the plugins and actually the dependency plugin manage that very well.

Closes #17.